requirements.txt
python-dotenv
openai
zxcvbn
requests
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
Flask-CORS
python-dotenv